git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/DJob/trunk@5398 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ DistribJob/lib/perl/Task.pm
@@ -120,8 +120,14 @@ sub passSubTask {
     chomp $date;
     print "\nNode: ".$node->getNum()." [$date] subTask $subTaskNum succeeded...".$subTask->getRunningTime()." seconds\n";
 
-    $self->integrateSubTaskResults($subTaskNum, $node, $nodeSlotDir,
-				   $self->{mainResultDir});
+    my $integRes = $self->integrateSubTaskResults($subTaskNum, $node, $nodeSlotDir,
+                                                  $self->{mainResultDir});
+    ##default is for $integRes to be null if command succeeded ... if returns 1 then want to fail subtask
+    if($integRes == 1){
+      $self->failSubTask($subTask);
+      return;
+    }
+
     &runCmd("/bin/rm -rf $subTaskDir");
     &log($self->{completeLog}, "$subTaskNum\n");
 }
